Webpack-Plugin for extracting the translation-ids
Prototype that is able to extract some translation id's with the default texts from html files.
The extraction is not 100% angular-translate conform and needs further cleaning.
Support for the filter is required.

The implementation does not support expressions for id's or default-texts.
The plugin does not support calls to the translation service because those calls are always dynamic
 -> Developer should create the translation entery himself.