Fix issue with webpack-karma
Fix an issue where webpack-karma fails when using the plugin.
The problem seams to be that webpack-karma does not expect to have
any assets when no file have been included (yet). In this case
webpack-karma failes with a not helping error message.
To avoid the issue, suppress adding the asset if no translations
have been found (seams leggit anyway).